requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
joblib==1.3.1
sentence-transformers==2.2.2
python-docx==0.8.11
PyPDF2==3.0.1
xlrd==2.0.1
python-dotenv==1.0.0
setuptools==68.0.0
# use conda for installing these packages
# pandas==1.5.3
# scikit-learn==1.3.1
# openai==0.28.0
